heroku-bgremover/requirements.txt
--extra-index-url https://google-coral.github.io/py-repo/
tflite_runtime
Flask
pillow
numpy
gunicorn